Small update.
Updated our keywords, operators, and separators to include what the Grader said we should have. Also included const int variables to store the sizes of those arrays so that we can just change those (if we ever need to) and so we don't have to look for the while loops that uses those values.